Finicial AI Scanner/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
Pillow==10.2.0
groq==0.4.2
pandas==2.2.1
numpy==1.26.4
plotly==5.18.0
pytesseract==0.3.10 